requirements.txt
Flask==3.1.3
mysql-connector-python==9.6.0
bcrypt==5.0.0
python-dotenv==1.2.2
opencv-python-headless
gunicorn
Jinja2==3.1.6
Werkzeug==3.1.6
itsdangerous==2.2.0
click==8.3.1
MarkupSafe==3.0.3
blinker==1.9.0